Data proc funciton updates
